LayoutTests/imported/w3c/web-platform-tests/content-security-policy/inheritance/blob-inherits-from-meta-http-equiv-with-invalid-characters-expected.txt

ALERT: PASS executed blob URL script.

PASS blob: URL inherits CSP from a meta tag whose contents have newline characters.
